add top navbar
