--- src/api/getAdminCohorts.tsx
@@ -1,5 +1,0 @@
-import api from './api';
-
-export const getAdminCohorts = () => {
-  return api.get('/admin/cohorts');
-};
