+++ src/api/getPlans.tsx
@@ -1,0 +1,6 @@
+import { CohortInterface } from '../types';
+import api from './api';
+
+export const getPlans = () => {
+  return api.get<CohortInterface[]>('/admin/cohorts');
+};
