+++ src/components/Navbar/Navbar.tsx
@@ -1,15 +1,28 @@
-import React, { ReactNode } from 'react';
-import { NavbarContainer, MenuItems } from './Navbar.styled';
+import React from 'react';
+import { useNavigate } from 'react-router-dom';
+import { logout } from '../../api/logout';
+import MenuItem from '../Sidebar/MenuItem';
+import { LogoutIcon } from '../Sidebar/Sidebar.styled';
+import { StyledNavbar } from './Navbar.styled';
 
-interface NavbarProps {
-  children: ReactNode;
-}
+const Navbar = () => {
+  const navigate = useNavigate();
 
-const Navbar = ({ children }: NavbarProps) => {
+  const handleLogout = () => {
+    logout();
+    navigate('/sign-in', { replace: true });
+  };
+
   return (
-    <NavbarContainer>
-      <MenuItems>{children}</MenuItems>
-    </NavbarContainer>
+    <StyledNavbar>
+      <div>Logo</div>
+      <MenuItem to="/member/home">Home</MenuItem>
+      <MenuItem to="/member/editor">Editor</MenuItem>
+      <MenuItem to="/sign-in" onClick={() => handleLogout()}>
+        <LogoutIcon />
+        Logout
+      </MenuItem>
+    </StyledNavbar>
   );
 };
