+++ src/pages/Home/Home.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { StyledHome } from './Home.styled';
+
+const Home = () => {
+  return <StyledHome>Home</StyledHome>;
+};
+
+export default Home;
