--- src/pages/Member/AdminDashboard.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const AdminDashboard = () => {
-  return <div>Admin dashboard</div>;
-};
-
-export default AdminDashboard;
