--- src/pages/Student/StudentClassTracker.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const StudentClassTracker = () => {
-  return <div>Student class tracker</div>;
-};
-
-export default StudentClassTracker;
