--- src/pages/Student/StudentDashboard.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const StudentDashboard = () => {
-  return <div>Student dashboard</div>;
-};
-
-export default StudentDashboard;
